refactor(runtime): split write_pipeline + lifecycle by concept; relocate stats_test
runtime/ becomes the small-helper layer with one file per concept:

  runtime/request.go         Request envelope (shared with engine/vlog)
  runtime/registry.go        Module + Registry generic pattern
  runtime/background.go      BackgroundServices orchestrator
  runtime/policy.go          (unchanged) hot-write + throttle helpers

The commit-pipeline-specific types (CommitRequest, CommitRequestPool,
CommitQueue, CommitBatch, SyncBatch) move into runtime/commit/types.go
alongside the Pipeline that owns them. runtime/commit/pipeline.go
drops the runtime.CommitX prefix on every internal reference and
keeps runtime.Request as the shared envelope only.

Tests follow the 1:1 source-test rule:
  runtime/request_test.go      Request lifecycle (Reset, Wait, DecrRef)
  runtime/registry_test.go     Module + Registry
  runtime/background_test.go   BackgroundServices lifecycle
  runtime/commit/types_test.go CommitQueue lifecycle + accounting
  runtime/stats/stats_test.go  Stats integration (moved from root,
                               external package iterator-style with
                               inline test fixtures)

Side cleanups:
- (db *DB).applyThrottle → ApplyThrottle (exported so the moved
  stats_test in runtime/stats can drive throttle state)
- db_test.go updated to use *commit.CommitRequest for the
  TestFinishCommitRequestsPerRequestErrors batch literal
